Handle continued number lists
Adds the ability to handle lists of the type:

Here comes a multi-part list.

1. Item the first.
2. Item the second.

Here is some intermediate text.

3. [@3] Item the third.
4. Item the fourth.

Previously the second list would start at 1. This change stores some
context during the line parsing to be able to continue the numbering at
3.

Also strip the [@3] field, it should not be part of the output.